Added Comments

+++ src/serverFunctions/Images/deleteImage.tsx
@@ -3,8 +3,11 @@
 import { prisma } from "@/db";
 import { unlinkSync } from "fs";
 import path from "path";
+import { requireAdmin } from "../user/requireAdmin";
 
 export async function deleteImage(filename: string) {
+  await requireAdmin();
+
   const image = await prisma.image.delete({
     where: {
       filename: filename,
